requirements.txt
click==6.6
coverage==4.2
coveralls==1.1
docopt==0.6.2
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dotenv==0.6.0
requests==2.11.1
Werkzeug==0.11.11
